requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.18
requests>=2.31.0
python-dotenv>=1.0.0

# AI and ML
google-generativeai>=0.3.0
scikit-learn>=1.3.0
ta-lib>=0.4.25
tensorflow>=2.13.0

# Data Processing
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scipy>=1.11.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Database
sqlalchemy>=2.0.0
sqlite3

# Trading and Finance
alpha-vantage>=2.3.0
finnhub-python>=2.4.0
marketstack>=0.1.0

# Utilities
pytz>=2023.3
python-dateutil>=2.8.0
schedule>=1.2.0
psutil>=5.9.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.0.0
black>=23.7.0
mypy>=1.5.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0

# Cloud Services
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0
google-cloud-logging>=3.8.0
